==> FINAL_folder/STEP5_streamlit_data_prep.ipynb <==
import pandas as pd
import ast

# Load the results DataFrame from the provided Excel file
file_path = '/content/drive/MyDrive/ALY6080/FINAL_folder/semantic_search_results_with_actual_score.xlsx'
df = pd.read_excel(file_path)
# Function to extract table names from the JSON-like string in the specified columns
def extract_table_names(table_info):
    try:
        table_dict = ast.literal_eval(table_info)
        table_names = ','.join(table_dict.keys())
        return table_names
    except (ValueError, SyntaxError):
        return ''

# Apply the function to create new columns 'Actual Table Names', 'Top 1 Table Names', etc.
df['Actual Table Names'] = df['Actual Table Info'].apply(extract_table_names)
df['Top 1 Table Names'] = df['Top 1 Similar Table Info'].apply(extract_table_names)
df['Top 2 Table Names'] = df['Top 2 Similar Table Info'].apply(extract_table_names)
df['Top 3 Table Names'] = df['Top 3 Similar Table Info'].apply(extract_table_names)

# Save the updated DataFrame to a new Excel file
output_path = '/content/drive/MyDrive/ALY6080/FINAL_folder/final_app.xlsx'
df.to_excel(output_path, index=False)
